Renamed loginService to authService.
